requirements.txt
streamlit
pillow
torch
torchaudio
torchvision